+++ test/reagent/impl/util_test.cljs
@@ -2,6 +2,33 @@
   (:require [clojure.test :refer [deftest is testing]]
             [reagent.impl.util :as util]))
 
+(deftest class-names-test
+  (is (= nil
+         (util/class-names)
+         (util/class-names nil)
+         (util/class-names [])
+         (util/class-names nil [])))
+  (is (= "a b"
+         (util/class-names ["a" "b"])
+         (util/class-names "a" "b")))
+  (is (= "a"
+         (util/class-names :a)
+         (util/class-names [:a])
+         (util/class-names nil "a")
+         (util/class-names [] nil "a")))
+  (is (= "a b c d"
+         (util/class-names "a" "b" nil ["c" "d"]))))
+
+; (simple-benchmark []
+;                   (do (util/class-names "a" "b")
+;                       (util/class-names nil "a")
+;                       (util/class-names "a" nil))
+;                   10000)
+
+; (simple-benchmark []
+;                   (util/class-names "a" "b" nil "c" "d")
+;                   10000)
+
 (deftest merge-props-test
   (testing "no arguments"
     (is (nil? (util/merge-props))))
@@ -36,9 +63,9 @@
                               :class "quux"}
                              nil))))
 
-  (testing ":class collection"
+  (testing ":class"
     (is (= {:class "foo bar baz quux"}
            (util/merge-props {:class "foo bar"}
                              {:class ["baz" "quux"]})
-           (util/merge-props {} {:class ["foo" "bar" "baz" "quux"]})
+           (util/merge-props nil {:class ["foo" "bar" "baz" "quux"]})
            (util/merge-props {:class ["foo" "bar" "baz" "quux"]})))))